Added Retrofit as a new RemoteDataSource.
Added RetrofitModule to build the retrofit instance via dagger2 which provides "BaseUrl, Gson, GsonConverterFactory and Retrofit".
Added RetrofitServicesModule to build each needed service using a retrofit instance (Which was provided via RetrofitModule).

The target was to inject the services into RemoteDataSourceUsingRetrofit class, which will in turn use the injected services to be provide the need data which will be requested by DataRepository.
To Added more services just add the service class, then added to RetrofitServicesModule, and added a constructor param to RemoteDataSourceUsingRetrofit. :)